Update to v093r08 release.
byuu says:

Changelog:
- Game Boy and Game Boy Color now have a weak hipass filter to remove DC
  bias (or whatever)
- Game Boy and Game Boy Color now have cycle-based PPU renderers instead
  of scanline-based renderers
- improved Game Boy color emulation palette contrast
- fixed GTK+ ListView selection bug
- fixed a typo when saving states (should say "Saved to slot N", not
  "Save to slot N")

+++ gb/video/video.hpp
@@ -7,7 +7,7 @@ struct Video {
 
 private:
   bool color_emulation;
-  static const double monochrome[4][3];
+  static const uint16 monochrome[4][3];
   uint32_t palette_dmg(unsigned color) const;
   uint32_t palette_sgb(unsigned color) const;
   uint32_t palette_cgb(unsigned color) const;
